Release

+++ src/main/java/com/conveyal/datatools/editor/jobs/ExportSnapshotToGTFSJob.java
@@ -21,11 +21,13 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
     private static final Logger LOG = LoggerFactory.getLogger(ExportSnapshotToGTFSJob.class);
     private final Snapshot snapshot;
     private final String feedVersionId;
+    private File tempFile;
 
     public ExportSnapshotToGTFSJob(Auth0UserProfile owner, Snapshot snapshot, String feedVersionId) {
         super(owner, "Exporting snapshot " + snapshot.name, JobType.EXPORT_SNAPSHOT_TO_GTFS);
         this.snapshot = snapshot;
         this.feedVersionId = feedVersionId;
+        status.update("Starting database snapshot...", 10);
     }
 
     public ExportSnapshotToGTFSJob(Auth0UserProfile owner, Snapshot snapshot) {
@@ -39,7 +41,6 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
 
     @Override
     public void jobLogic() {
-        File tempFile;
         try {
             tempFile = File.createTempFile("snapshot", "zip");
         } catch (IOException e) {
@@ -50,9 +51,8 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
         JdbcGtfsExporter exporter = new JdbcGtfsExporter(snapshot.namespace, tempFile.getAbsolutePath(), DataManager.GTFS_DATA_SOURCE, true);
         FeedLoadResult result = exporter.exportTables();
         if (result.fatalException != null) {
-            String message = String.format("Error (%s) encountered while exporting database tables.", result.fatalException);
-            LOG.error(message);
-            status.fail(message);
+            status.fail(String.format("Error (%s) encountered while exporting database tables.", result.fatalException));
+            return;
         }
 
         // Override snapshot ID if exporting feed for use as new feed version.
@@ -60,6 +60,7 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
         String bucketPrefix = feedVersionId != null ? "gtfs" : "snapshots";
         // FIXME: replace with use of refactored FeedStore.
         // Store the project merged zip locally or on s3
+        status.update("Writing snapshot to GTFS file", 90);
         if (DataManager.useS3) {
             String s3Key = String.format("%s/%s", bucketPrefix, filename);
             FeedStore.s3Client.putObject(DataManager.feedBucket, s3Key, tempFile);
@@ -68,12 +69,18 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
             try {
                 FeedVersion.feedStore.newFeed(filename, new FileInputStream(tempFile), null);
             } catch (IOException e) {
-                LOG.error("Could not store feed for snapshot {}", snapshot.id);
-                e.printStackTrace();
-                status.fail("Could not export snapshot to GTFS.");
+                status.fail(String.format("Could not store feed for snapshot %s", snapshot.id), e);
             }
         }
+    }
+
+    @Override
+    public void jobFinished () {
+        if (!status.error) status.completeSuccessfully("Export complete!");
         // Delete snapshot temp file.
-        tempFile.delete();
+        if (tempFile != null) {
+            LOG.info("Deleting temporary GTFS file for exported snapshot at {}", tempFile.getAbsolutePath());
+            tempFile.delete();
+        }
     }
 }
